+++ lambdas/common/granule_tracker.py
